+++ src/css/selectors.rs
@@ -6,10 +6,12 @@
 //!
 //! This _may_ change to a something like a tree of operators with
 //! leafs of simple selectors in some future release.
-use crate::css::Value;
-use crate::sass::SassString;
+use super::{is_not, CssString, Value};
+use crate::parser::css::{selector, selector_part, selectors};
+use crate::parser::input_span;
+use crate::parser::{ParseError, SourcePos};
 use crate::value::ListSeparator;
-use crate::{Error, ParseError, ScopeRef};
+use std::convert::{TryFrom, TryInto};
 use std::fmt;
 use std::io::Write;
 
@@ -17,7 +19,7 @@ use std::io::Write;
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd)]
 pub struct Selectors {
     /// The actual selectors.
-    pub s: Vec<Selector>,
+    s: Vec<Selector>,
     backref: Selector,
 }
 
@@ -33,16 +35,54 @@ impl Selectors {
     /// Create a new Selectors from a vec of selectors.
     pub fn new(s: Vec<Selector>) -> Self {
         Selectors {
-            s,
+            s: if s.is_empty() {
+                vec![Selector::root()]
+            } else {
+                s
+            },
             backref: Selector::root(),
         }
     }
+    /// Validate that this selector is ok to use in css.
+    ///
+    /// `Selectors` can contain backref (`&`), but those must be
+    /// resolved before using the `Selectors` in css.
+    pub fn css_ok(self) -> Result<Self, BadSelector> {
+        if self.has_backref() {
+            let sel = self.to_string();
+            Err(BadSelector::Backref(input_span(sel.as_bytes()).into()))
+        } else {
+            Ok(self)
+        }
+    }
+
     /// Remove the first of these selectors (or the root selector if empty).
-    pub fn one(&self) -> Selector {
+    pub(crate) fn one(&self) -> Selector {
         self.s.first().cloned().unwrap_or_else(Selector::root)
     }
+
+    pub(crate) fn append(self, ext: Self) -> Result<Self, crate::Error> {
+        let ext = ext.css_ok()?;
+        Ok(Selectors::new(
+            self.s
+                .into_iter()
+                .flat_map(|b| {
+                    ext.s.iter().map(move |e| {
+                        if e.0[0].is_operator() || e.0[0].is_wildcard() {
+                            return Err(crate::Error::error(&format!(
+                                "Can't append {} to {}.",
+                                e, b
+                            )));
+                        }
+                        parse_selector(&format!("{}{}", b, e))
+                    })
+                })
+                .collect::<Result<_, _>>()?,
+        ))
+    }
+
     /// Create the full selector for when self is used inside a parent selector.
-    pub fn inside(&self, parent: &Self) -> Self {
+    pub(crate) fn inside(&self, parent: &Self) -> Self {
         let mut result = Vec::new();
         for p in &parent.s {
             for s in &self.s {
@@ -54,21 +94,34 @@ impl Selectors {
             backref: parent.backref.clone(),
         }
     }
+
+    /// True if any of the selectors contains a backref (`&`).
+    pub(crate) fn has_backref(&self) -> bool {
+        self.s.iter().any(|s| s.has_backref())
+    }
+
     /// Get these selectors with a specific backref selector.
     ///
     /// Used to create `@at-root` contexts, to have `&` work in them.
-    pub fn with_backref(self, context: Selector) -> Self {
+    pub(crate) fn with_backref(self, context: Selector) -> Self {
         self.inside(&Selectors {
             s: vec![Selector::root()],
             backref: context,
         })
     }
-    /// Create a sass `Value` representing this set of selectors.
-    pub fn to_value(&self) -> Value {
-        if self.s.len() == 1 && self.s[0].0.is_empty() {
+}
+
+impl From<Selectors> for Value {
+    /// Create a css `Value` representing a set of selectors.
+    ///
+    /// The result will be a comma-separated [list](Value::List) of
+    /// space-separated lists of strings, or [null](Value::Null) if
+    /// this is a root (empty) selector.
+    fn from(sel: Selectors) -> Value {
+        if sel.is_root() {
             return Value::Null;
         }
-        let content = self
+        let content = sel
             .s
             .iter()
             .map(|s: &Selector| {
@@ -105,24 +158,94 @@ impl Selectors {
             .collect::<Vec<_>>();
         Value::List(content, Some(ListSeparator::Comma), false)
     }
-    /// Evaluate any interpolation in these Selectors.
-    pub fn eval(&self, scope: ScopeRef) -> Result<Selectors, Error> {
-        let s = Selectors::new(
-            self.s
-                .iter()
-                .map(|s| s.eval(scope.clone()))
-                .collect::<Result<Vec<_>, Error>>()?,
-        );
-        // The "simple" parts we get from evaluating interpolations may
-        // contain high-level selector separators (i.e. ","), so we need to
-        // parse the selectors again, from a string representation.
-        use crate::parser::input_span;
-        use crate::parser::selectors::selectors;
-        // TODO: Get the span from the source of self!
-        Ok(ParseError::check(selectors(input_span(
-            format!("{} ", s).as_bytes(),
-        )))?)
+}
+
+impl TryFrom<Value> for Selectors {
+    type Error = BadSelector;
+    fn try_from(v: Value) -> Result<Self, Self::Error> {
+        value_to_selectors(&v).map_err(move |e| e.ctx(v))
     }
+}
+fn value_to_selectors(v: &Value) -> Result<Selectors, BadSelector0> {
+    match v {
+        Value::List(vv, s, _) => match s {
+            Some(ListSeparator::Comma) => {
+                let vv = vv
+                    .iter()
+                    .map(value_to_selector)
+                    .collect::<Result<_, _>>()?;
+                Ok(Selectors::new(vv))
+            }
+            Some(ListSeparator::Space) => {
+                let (mut outer, last) = vv.iter().fold(
+                    Result::<_, BadSelector0>::Ok((vec![], vec![])),
+                    |a, v: &Value| {
+                        let (mut outer, mut a) = a?;
+                        if let Ok(ref mut s) = check_selector_str(v) {
+                            push_descendant(&mut a, s)
+                        } else {
+                            let mut s = parse_selectors_str(v)?;
+                            if let Some(f) = s.s.first_mut() {
+                                push_descendant(&mut a, f);
+                                std::mem::swap(&mut a, &mut f.0);
+                            }
+                            if let Some(last) = s.s.pop() {
+                                a = last.0;
+                            }
+                            outer.extend(s.s);
+                        }
+                        Ok((outer, a))
+                    },
+                )?;
+                outer.push(Selector(last));
+                Ok(Selectors::new(outer))
+            }
+            _ => Err(BadSelector0::Value),
+        },
+        Value::Literal(s) => {
+            if s.value().is_empty() {
+                Ok(Selectors::root())
+            } else {
+                let span = input_span(s.value().as_bytes());
+                Ok(ParseError::check(selectors(span))?)
+            }
+        }
+        _ => Err(BadSelector0::Value),
+    }
+}
+
+fn check_selector_str(v: &Value) -> Result<Selector, BadSelector0> {
+    match v {
+        Value::Literal(s) => {
+            if s.value().is_empty() {
+                Ok(Selector::root())
+            } else {
+                let span = input_span(s.value().as_bytes());
+                Ok(ParseError::check(selector(span))?)
+            }
+        }
+        _ => Err(BadSelector0::Value),
+    }
+}
+fn parse_selectors_str(v: &Value) -> Result<Selectors, BadSelector0> {
+    match v {
+        Value::Literal(s) => {
+            if s.value().is_empty() {
+                Ok(Selectors::root())
+            } else {
+                let span = input_span(s.value().as_bytes());
+                Ok(ParseError::check(selectors(span))?)
+            }
+        }
+        _ => Err(BadSelector0::Value),
+    }
+}
+
+fn push_descendant(to: &mut Vec<SelectorPart>, from: &mut Selector) {
+    if !to.is_empty() {
+        to.push(SelectorPart::Descendant)
+    }
+    to.append(&mut from.0);
 }
 
 /// A css (or sass) selector.
@@ -130,7 +253,7 @@ impl Selectors {
 /// A selector does not contain `,`.  If it does, it is a `Selectors`,
 /// where each of the parts separated by the comma is a `Selector`.
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd)]
-pub struct Selector(pub Vec<SelectorPart>);
+pub struct Selector(pub(crate) Vec<SelectorPart>);
 
 impl Selector {
     /// Get the root (empty) selector.
@@ -158,27 +281,56 @@ impl Selector {
         }
     }
 
-    fn eval(&self, scope: ScopeRef) -> Result<Selector, Error> {
-        self.0
-            .iter()
-            .map(|sp| sp.eval(scope.clone()))
-            .collect::<Result<_, _>>()
-            .map(Selector)
-    }
-
     fn has_backref(&self) -> bool {
         self.0.iter().any(|p| p.has_backref())
     }
+}
+
+impl TryInto<Selector> for Value {
+    type Error = BadSelector;
+    fn try_into(self) -> Result<Selector, Self::Error> {
+        value_to_selector(&self).map_err(move |e| e.ctx(self))
+    }
+}
+// Internal, the api is try_into.
+fn value_to_selector(v: &Value) -> Result<Selector, BadSelector0> {
+    match v {
+        Value::List(list, sep, _)
+            if sep == &None || sep == &Some(ListSeparator::Space) =>
+        {
+            list_to_selector(list)
+        }
+        Value::Literal(s) => {
+            if s.value().is_empty() {
+                Ok(Selector::root())
+            } else {
+                let span = input_span(s.value().as_bytes());
+                Ok(ParseError::check(selector(span))?)
+            }
+        }
+        _ => Err(BadSelector0::Value),
+    }
+}
+
+fn list_to_selector(list: &[Value]) -> Result<Selector, BadSelector0> {
+    Ok(Selector(list.iter().fold(
+        Result::<_, BadSelector0>::Ok(vec![]),
+        |a, v| {
+            let mut a = a?;
+            if !a.is_empty() {
+                a.push(SelectorPart::Descendant)
+            }
+            a.push(value_to_selector_part(v)?);
+            Ok(a)
+        },
+    )?))
 }
 
 /// A selector consist of a sequence of these parts.
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd)]
 pub enum SelectorPart {
     /// A simple selector, eg a class, id or element name.
-    ///
-    /// Note that a Simple selector can hide a more complex selector
-    /// through string interpolation.
-    Simple(SassString),
+    Simple(String),
     /// The empty relational operator.
     ///
     /// The thing after this is a descendant of the thing before this.
@@ -188,25 +340,25 @@ pub enum SelectorPart {
     /// An attribute selector
     Attribute {
         /// The attribute name
-        name: SassString,
+        name: CssString,
         /// An operator
         op: String,
         /// A value to match.
-        val: SassString,
+        val: CssString,
         /// Optional modifier.
         modifier: Option<char>,
     },
     /// A css3 pseudo-element (::foo)
     PseudoElement {
         /// The name of the pseudo-element
-        name: SassString,
+        name: CssString,
         /// Arguments to the pseudo-element
         arg: Option<Selectors>,
     },
     /// A pseudo-class or a css2 pseudo-element (:foo)
     Pseudo {
         /// The name of the pseudo-class
-        name: SassString,
+        name: CssString,
         /// Arguments to the pseudo-class
         arg: Option<Selectors>,
     },
@@ -215,7 +367,7 @@ pub enum SelectorPart {
 }
 
 impl SelectorPart {
-    fn is_operator(&self) -> bool {
+    pub(crate) fn is_operator(&self) -> bool {
         match *self {
             SelectorPart::Descendant | SelectorPart::RelOp(_) => true,
             SelectorPart::Simple(_)
@@ -223,6 +375,13 @@ impl SelectorPart {
             | SelectorPart::PseudoElement { .. }
             | SelectorPart::Pseudo { .. }
             | SelectorPart::BackRef => false,
+        }
+    }
+    pub(crate) fn is_wildcard(&self) -> bool {
+        if let SelectorPart::Simple(s) = self {
+            s == "*"
+        } else {
+            false
         }
     }
     fn has_backref(&self) -> bool {
@@ -268,44 +427,20 @@ impl SelectorPart {
             }
         }
     }
-    fn eval(&self, scope: ScopeRef) -> Result<SelectorPart, Error> {
-        match *self {
-            SelectorPart::Attribute {
-                ref name,
-                ref op,
-                ref val,
-                ref modifier,
-            } => Ok(SelectorPart::Attribute {
-                name: name.evaluate(scope.clone())?.into(),
-                op: op.clone(),
-                val: val.evaluate(scope)?.opt_unquote().into(),
-                modifier: *modifier,
-            }),
-            SelectorPart::Simple(ref v) => {
-                Ok(SelectorPart::Simple(v.evaluate(scope)?.into()))
-            }
-            SelectorPart::Pseudo { ref name, ref arg } => {
-                let arg = match &arg {
-                    Some(ref a) => Some(a.eval(scope.clone())?),
-                    None => None,
-                };
-                Ok(SelectorPart::Pseudo {
-                    name: name.evaluate(scope)?.into(),
-                    arg,
-                })
-            }
-            SelectorPart::PseudoElement { ref name, ref arg } => {
-                let arg = match &arg {
-                    Some(ref a) => Some(a.eval(scope.clone())?),
-                    None => None,
-                };
-                Ok(SelectorPart::PseudoElement {
-                    name: name.evaluate(scope)?.into(),
-                    arg,
-                })
-            }
-            ref sp => Ok(sp.clone()),
-        }
+}
+
+impl TryInto<SelectorPart> for Value {
+    type Error = BadSelector;
+    fn try_into(self) -> Result<SelectorPart, Self::Error> {
+        value_to_selector_part(&self).map_err(move |e| e.ctx(self))
+    }
+}
+fn value_to_selector_part(v: &Value) -> Result<SelectorPart, BadSelector0> {
+    match v {
+        Value::Literal(s) => Ok(ParseError::check(selector_part(
+            input_span(s.value().as_bytes()),
+        ))?),
+        _ => Err(BadSelector0::Value),
     }
 }
 
@@ -437,4 +572,66 @@ mod test {
         );
         assert_eq!(format!("{}", s), "foo.bar")
     }
+}
+
+enum BadSelector0 {
+    Value,
+    Parse(ParseError),
+}
+impl BadSelector0 {
+    fn ctx(self, v: Value) -> BadSelector {
+        match self {
+            Self::Value => BadSelector::Value(v),
+            Self::Parse(err) => BadSelector::Parse(err),
+        }
+    }
+}
+impl From<ParseError> for BadSelector0 {
+    fn from(e: ParseError) -> Self {
+        BadSelector0::Parse(e)
+    }
+}
+
+/// The error when a [Value] cannot be converted to a [Selectors] or [Selector].
+#[derive(Debug)]
+pub enum BadSelector {
+    /// The value was not the expected type of list or string.
+    Value(Value),
+    /// There was an error parsing a string value.
+    Parse(ParseError),
+    /// A backref (`&`) were present but not allowed there.
+    Backref(SourcePos),
+}
+
+impl fmt::Display for BadSelector {
+    fn fmt(&self, out: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            BadSelector::Value(v) => out.write_str(&is_not(
+                v,
+                "a valid selector: it must be a string,\
+                 \na list of strings, or a list of lists of strings",
+            )),
+            BadSelector::Parse(e) => e.fmt(out),
+            BadSelector::Backref(pos) => {
+                writeln!(out, "Parent selectors aren\'t allowed here.")?;
+                pos.show(out)
+            }
+        }
+    }
+}
+impl From<BadSelector> for crate::Error {
+    fn from(e: BadSelector) -> crate::Error {
+        match e {
+            BadSelector::Parse(e) => e.into(),
+            e => crate::Error::error(e.to_string()),
+        }
+    }
+}
+impl From<ParseError> for BadSelector {
+    fn from(e: ParseError) -> Self {
+        BadSelector::Parse(e)
+    }
+}
+fn parse_selector(s: &str) -> Result<Selector, crate::Error> {
+    Ok(ParseError::check(selector(input_span(s.as_bytes())))?)
 }